+++ lab2/mystrlen.h
@@ -1,4 +1,4 @@
 #ifndef MYSTRLEN_H_
 #define MYSTRLEN_H_
-    unsigned long long mystrlen(const char *s);
+    size_t mystrlen(const char *s);
 #endif
